+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/toolWindow/components/ProcessListComponent.kt
@@ -1,40 +1,45 @@
 package com.github.rafaelldi.diagnosticsclientplugin.toolWindow.components
 
 import com.github.rafaelldi.diagnosticsclientplugin.generated.ProcessInfo
+import com.intellij.ide.CopyProvider
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.DataContext
+import com.intellij.openapi.ide.CopyPasteManager
 import com.intellij.ui.ColoredListCellRenderer
 import com.intellij.ui.ListSpeedSearch
+import com.intellij.ui.SortedListModel
 import com.intellij.ui.components.JBList
 import icons.RiderIcons
-import javax.swing.DefaultListModel
+import java.awt.datatransfer.StringSelection
 import javax.swing.JList
-import javax.swing.tree.TreeSelectionModel
+import javax.swing.ListSelectionModel
 
-class ProcessListComponent : JBList<ProcessInfo>() {
-    private val listModel = DefaultListModel<ProcessInfo>()
+class ProcessListComponent : JBList<Pair<Int, ProcessInfo>>(), CopyProvider {
+    private val listModel = SortedListModel<Pair<Int, ProcessInfo>>(compareBy { it.first })
 
     init {
         model = listModel
         emptyText.text = "No processes to display"
-        selectionModel.selectionMode = TreeSelectionModel.SINGLE_TREE_SELECTION
+        selectionMode = ListSelectionModel.SINGLE_SELECTION
         cellRenderer = CellRenderer()
-        ListSpeedSearch(this)
+        ListSpeedSearch(this) { it.second.processName }
     }
 
     val selectedProcessId: Int?
-        get() = selectedValue?.processId
+        get() = selectedValue?.first
 
-    fun add(item: ProcessInfo) {
-        listModel.addElement(item)
+    fun add(pid: Int, item: ProcessInfo) {
+        listModel.add(pid to item)
     }
 
-    fun remove(item: ProcessInfo) {
-        listModel.removeElement(item)
+    fun remove(pid: Int, item: ProcessInfo) {
+        listModel.remove(pid to item)
     }
 
-    private class CellRenderer : ColoredListCellRenderer<ProcessInfo>() {
+    private class CellRenderer : ColoredListCellRenderer<Pair<Int, ProcessInfo>>() {
         override fun customizeCellRenderer(
-            list: JList<out ProcessInfo>,
-            value: ProcessInfo?,
+            list: JList<out Pair<Int, ProcessInfo>>,
+            value: Pair<Int, ProcessInfo>?,
             index: Int,
             selected: Boolean,
             hasFocus: Boolean
@@ -43,8 +48,22 @@ class ProcessListComponent : JBList<ProcessInfo>() {
                 return
             }
 
-            append("${value.processId} - ${value.processName}")
+            append("${value.first} - ${value.second.processName}")
             icon = RiderIcons.RunConfigurations.DotNetProject
         }
     }
+
+    override fun performCopy(dataContext: DataContext) {
+        if (!isSelectionEmpty) {
+            CopyPasteManager.getInstance().setContents(StringSelection(getTextForCopy()))
+        }
+    }
+
+    private fun getTextForCopy() = "${selectedValue.first} - ${selectedValue.second.processName}"
+
+    override fun isCopyEnabled(dataContext: DataContext) = !isSelectionEmpty
+
+    override fun isCopyVisible(dataContext: DataContext) = !isSelectionEmpty
+
+    override fun getActionUpdateThread() = ActionUpdateThread.BGT
 }